+++ app/src/main/java/ch/epfl/polybazaar/widgets/permissions/PermissionUtils.java
@@ -1,38 +1,47 @@
 package ch.epfl.polybazaar.widgets.permissions;
 
-
-import android.Manifest;
-import android.app.AlertDialog;
-import android.app.Dialog;
-import android.content.DialogInterface;
 import android.content.pm.PackageManager;
-import android.os.Bundle;
 
+import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.core.app.ActivityCompat;
-import androidx.fragment.app.DialogFragment;
-import android.widget.Toast;
+import androidx.core.content.ContextCompat;
 
-import ch.epfl.polybazaar.R;
+import android.util.Log;
 
 /**
  * Utility class for access to runtime permissions.
  */
 public abstract class PermissionUtils {
 
-    /**
-     * Requests the fine location permission. If a rationale with an additional explanation should
-     * be shown to the user, displays a dialog that triggers the request.
-     */
-    public static void requestPermission(AppCompatActivity activity, int requestId,
-                                         String permission, boolean finishActivity) {
-        if (ActivityCompat.shouldShowRequestPermissionRationale(activity, permission)) {
+    private static final String TAG = "PermissionUtils";
+
+    public static boolean assertPermission(AppCompatActivity currentContext, @NonNull String permission) {
+        if (isPermissionGranted(currentContext,permission)) {
+            Log.d(TAG, "Permission " + permission + " already granted");
+            return true;
+        } else {
+            Log.d(TAG, "Permission " + permission + " not yet granted");
+            requestPermission(currentContext, permission);
+            return isPermissionGranted(currentContext,permission);
+        }
+    }
+
+    private static boolean isPermissionGranted(AppCompatActivity currentContext, @NonNull String permission) {
+        return ContextCompat.checkSelfPermission(currentContext, "android.Manifest.permission." + permission)
+                == PackageManager.PERMISSION_GRANTED;
+    }
+
+    private static void requestPermission(AppCompatActivity currentContext, String permission) {
+        int ID = (int)(Math.random()*100000.0);
+        if (ActivityCompat.shouldShowRequestPermissionRationale(currentContext, "android.Manifest.permission." + permission)) {
             // Display a dialog with rationale.
-            PermissionUtils.RationaleDialog.newInstance(requestId, finishActivity)
-                    .show(activity.getSupportFragmentManager(), "dialog");
+            // Here we tell the user why wee need to have this permission
+            RationaleDialog.newInstance(ID, false)
+                    .show(currentContext.getSupportFragmentManager(), "dialog");
         } else {
             // Location permission has not been granted yet, request it.
-            ActivityCompat.requestPermissions(activity, new String[]{permission}, requestId);
+            ActivityCompat.requestPermissions(currentContext, new String[]{"android.Manifest.permission." + permission}, ID);
 
         }
     }
@@ -53,119 +62,5 @@ public abstract class PermissionUtils {
         return false;
     }
 
-    /**
-     * A dialog that displays a permission denied message.
-     */
-    public static class PermissionDeniedDialog extends DialogFragment {
 
-        private static final String ARGUMENT_FINISH_ACTIVITY = "finish";
-
-        private boolean mFinishActivity = false;
-
-        /**
-         * Creates a new instance of this dialog and optionally finishes the calling Activity
-         * when the 'Ok' button is clicked.
-         */
-        public static PermissionDeniedDialog newInstance(boolean finishActivity) {
-            Bundle arguments = new Bundle();
-            arguments.putBoolean(ARGUMENT_FINISH_ACTIVITY, finishActivity);
-
-            PermissionDeniedDialog dialog = new PermissionDeniedDialog();
-            dialog.setArguments(arguments);
-            return dialog;
-        }
-
-        @Override
-        public Dialog onCreateDialog(Bundle savedInstanceState) {
-            mFinishActivity = getArguments().getBoolean(ARGUMENT_FINISH_ACTIVITY);
-
-            return new AlertDialog.Builder(getActivity())
-                    .setMessage(R.string.location_permission_denied)
-                    .setPositiveButton(android.R.string.ok, null)
-                    .create();
-        }
-
-        @Override
-        public void onDismiss(DialogInterface dialog) {
-            super.onDismiss(dialog);
-            if (mFinishActivity) {
-                Toast.makeText(getActivity(), R.string.permission_required_toast,
-                        Toast.LENGTH_SHORT).show();
-                getActivity().finish();
-            }
-        }
-    }
-
-    /**
-     * A dialog that explains the use of the location permission and requests the necessary
-     * permission.
-     * <p>
-     * The activity should implement
-     * {@link androidx.core.app.ActivityCompat.OnRequestPermissionsResultCallback}
-     * to handle permit or denial of this permission request.
-     */
-    public static class RationaleDialog extends DialogFragment {
-
-        private static final String ARGUMENT_PERMISSION_REQUEST_CODE = "requestCode";
-
-        private static final String ARGUMENT_FINISH_ACTIVITY = "finish";
-
-        private boolean mFinishActivity = false;
-
-        /**
-         * Creates a new instance of a dialog displaying the rationale for the use of the location
-         * permission.
-         * <p>
-         * The permission is requested after clicking 'ok'.
-         *
-         * @param requestCode    Id of the request that is used to request the permission. It is
-         *                       returned to the
-         *                       {@link androidx.core.app.ActivityCompat.OnRequestPermissionsResultCallback}.
-         * @param finishActivity Whether the calling Activity should be finished if the dialog is
-         *                       cancelled.
-         */
-        public static RationaleDialog newInstance(int requestCode, boolean finishActivity) {
-            Bundle arguments = new Bundle();
-            arguments.putInt(ARGUMENT_PERMISSION_REQUEST_CODE, requestCode);
-            arguments.putBoolean(ARGUMENT_FINISH_ACTIVITY, finishActivity);
-            RationaleDialog dialog = new RationaleDialog();
-            dialog.setArguments(arguments);
-            return dialog;
-        }
-
-        @Override
-        public Dialog onCreateDialog(Bundle savedInstanceState) {
-            Bundle arguments = getArguments();
-            final int requestCode = arguments.getInt(ARGUMENT_PERMISSION_REQUEST_CODE);
-            mFinishActivity = arguments.getBoolean(ARGUMENT_FINISH_ACTIVITY);
-
-            return new AlertDialog.Builder(getActivity())
-                    .setMessage(R.string.permission_rationale_location)
-                    .setPositiveButton(android.R.string.ok, new DialogInterface.OnClickListener() {
-                        @Override
-                        public void onClick(DialogInterface dialog, int which) {
-                            // After click on Ok, request the permission.
-                            ActivityCompat.requestPermissions(getActivity(),
-                                    new String[]{Manifest.permission.ACCESS_FINE_LOCATION},
-                                    requestCode);
-                            // Do not finish the Activity while requesting permission.
-                            mFinishActivity = false;
-                        }
-                    })
-                    .setNegativeButton(android.R.string.cancel, null)
-                    .create();
-        }
-
-        @Override
-        public void onDismiss(DialogInterface dialog) {
-            super.onDismiss(dialog);
-            if (mFinishActivity) {
-                Toast.makeText(getActivity(),
-                        R.string.permission_required_toast,
-                        Toast.LENGTH_SHORT)
-                        .show();
-                getActivity().finish();
-            }
-        }
-    }
 }
