--- app/src/main/java/ch/epfl/polybazaar/widgets/permissions/Permissions.java
@@ -1,5 +1,0 @@
-package ch.epfl.polybazaar.widgets.permissions;
-
-public class Permissions {
-    public static final int LOCATION_PERMISSION_REQUEST_CODE = 0;
-}
