+++ app/src/main/java/ch/epfl/polybazaar/widgets/permissions/RationaleDialog.java
@@ -1,0 +1,87 @@
+package ch.epfl.polybazaar.widgets.permissions;
+
+
+import android.Manifest;
+import android.app.AlertDialog;
+import android.app.Dialog;
+import android.content.DialogInterface;
+import android.os.Bundle;
+import android.widget.Toast;
+
+import androidx.core.app.ActivityCompat;
+import androidx.fragment.app.DialogFragment;
+
+import ch.epfl.polybazaar.R;
+
+/**
+ * A dialog that explains the use of the location permission and requests the necessary
+ * permission.
+ * <p>
+ * The activity should implement
+ * {@link androidx.core.app.ActivityCompat.OnRequestPermissionsResultCallback}
+ * to handle permit or denial of this permission request.
+ */
+public class RationaleDialog extends DialogFragment {
+
+    private static final String ARGUMENT_PERMISSION_REQUEST_CODE = "requestCode";
+
+    private static final String ARGUMENT_FINISH_ACTIVITY = "finish";
+
+    private boolean mFinishActivity = false;
+
+    /**
+     * Creates a new instance of a dialog displaying the rationale for the use of the location
+     * permission.
+     * <p>
+     * The permission is requested after clicking 'ok'.
+     *
+     * @param requestCode    Id of the request that is used to request the permission. It is
+     *                       returned to the
+     *                       {@link androidx.core.app.ActivityCompat.OnRequestPermissionsResultCallback}.
+     * @param finishActivity Whether the calling Activity should be finished if the dialog is
+     *                       cancelled.
+     */
+    public static RationaleDialog newInstance(int requestCode, boolean finishActivity) {
+        Bundle arguments = new Bundle();
+        arguments.putInt(ARGUMENT_PERMISSION_REQUEST_CODE, requestCode);
+        arguments.putBoolean(ARGUMENT_FINISH_ACTIVITY, finishActivity);
+        RationaleDialog dialog = new RationaleDialog();
+        dialog.setArguments(arguments);
+        return dialog;
+    }
+
+    @Override
+    public Dialog onCreateDialog(Bundle savedInstanceState) {
+        Bundle arguments = getArguments();
+        final int requestCode = arguments.getInt(ARGUMENT_PERMISSION_REQUEST_CODE);
+        mFinishActivity = arguments.getBoolean(ARGUMENT_FINISH_ACTIVITY);
+
+        return new AlertDialog.Builder(getActivity())
+                .setMessage(R.string.permission_rationale_location)
+                .setPositiveButton(android.R.string.ok, new DialogInterface.OnClickListener() {
+                    @Override
+                    public void onClick(DialogInterface dialog, int which) {
+                        // After click on Ok, request the permission.
+                        ActivityCompat.requestPermissions(getActivity(),
+                                new String[]{Manifest.permission.ACCESS_FINE_LOCATION},
+                                requestCode);
+                        // Do not finish the Activity while requesting permission.
+                        mFinishActivity = false;
+                    }
+                })
+                .setNegativeButton(android.R.string.cancel, null)
+                .create();
+    }
+
+    @Override
+    public void onDismiss(DialogInterface dialog) {
+        super.onDismiss(dialog);
+        if (mFinishActivity) {
+            Toast.makeText(getActivity(),
+                    R.string.permission_required_toast,
+                    Toast.LENGTH_SHORT)
+                    .show();
+            getActivity().finish();
+        }
+    }
+}
